requirements.txt
Django==4.2.20
Pillow==11.1.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
pydicom==2.4.4
numpy==1.26.4 